fix: modify characters change timing
