+++ app/manifest.json
@@ -2,11 +2,15 @@
   "manifest_version": 3,
   "name": "Enhance GitHub Timestamps",
   "version": "1.1.5",
+  "permissions": ["tabs"],
   "content_scripts": [
     {
       "matches": ["https://github.com/*"],
-      "js": ["/src/contents/contents.ts"],
+      "js": ["/src/01_frameworks-driver.layer/contents/contents.ts"],
       "run_at": "document_end"
     }
-  ]
+  ],
+  "background": {
+    "service_worker": "/src/01_frameworks-driver.layer/background/background.ts"
+  }
 }
