+++ app/src/01_frameworks-driver.layer/background/background.ts
@@ -1,0 +1,5 @@
+chrome.tabs.onUpdated.addListener((tabId, changeInfo, tab) => {
+  if (!tab.url?.includes("https://github.com/")) return;
+  if (changeInfo.status !== "complete") return;
+  chrome.tabs.sendMessage(tabId, undefined).catch(() => {});
+});
